Fix For Plunger not starting from Home
The travel distance for the plunger in both the directions is set to stroke length.

There was a slight error that was accumulated overtime and the stroke was not starting from home.

Added an extra travel distance to exhale cycle to make sure we move the plunger back to home without accumulating any error.

A more permanent fix would be to sense home at the beginning of the inhale cycle and at the end of the exhale cycle.